+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [1.1.0] - 25/10/2019
 
 * Added the option to make sheetBelow and sheetAbove widget draggable.
-* Update the readme
+* Fix minor bugs with the snapping system.
   
 ## [1.0.0] - 18/10/2019